Moving Foreground!!!!
This build I introduced moving foreground to help create a parallax
effect, I had problems with two many objects being created and needed to
find a way to remove junk, I created a physics body to collide with the
other physics bodies near the back of the screen (need to change this to
relative positions) and remove physics bodies.

Also I removed the physics body from the mights and tights, it was not
needed and negatively effected game play.